Refactored dbConnection, create var DbConn that use as connection to DB for dbManager functions. In dbManager create func GetByBrand
